+++ conftest.py
@@ -1,0 +1,4 @@
+"""Configuration for tests."""
+pytest_plugins = [
+    "secmlt.tests.fixtures",
+]
